feat: 完善用户、角色、部门管理功能并添加文件上传功能

+++ admin-back/src/main/java/com/zcy/ias/service/SysUserService.java
@@ -4,7 +4,7 @@ import com.baomidou.mybatisplus.extension.service.IService;
 import com.zcy.ias.entity.SysUser;
 
 /**
- * 系统用户表(SysUser)表服务接口
+ * 用户表(SysUser)表服务接口
  *
  * @author zhuangchongyi
  * @since 2025-03-25 18:23:14
@@ -16,7 +16,7 @@ public interface SysUserService extends IService<SysUser> {
      *
      * @param userId 用户id
      */
-    void updateLoginInfo(Integer userId);
+    void updateLoginInfo(Long userId);
 
 }
